rearrange pages for fewer clicks to email sending
