+++ app/support/page.tsx
@@ -1,5 +1,7 @@
-export default function Support(){
+export default function Support() {
     return (
-        <h1>Not yet implemented. Please contact tom@esgnaturerisk.com regarding your concern and we'll do our best to meet your needs!</h1>
-    )
+    <h1>
+        {"Not yet implemented. Please contact tom@esgnaturerisk.com regarding your concern and we'll do our best to meet your needs!"}
+    </h1>
+  );
 }
